Adding Flash Messenger

+++ src/main/java/com/merlobranco/springboot/app/controllers/ClienteController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.bind.support.SessionStatus;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.merlobranco.springboot.app.models.entity.Cliente;
 import com.merlobranco.springboot.app.models.service.ClienteService;
@@ -38,32 +39,42 @@ public class ClienteController {
 	}
 	
 	@GetMapping("/form/{id}")
-	public String editar(@PathVariable(value="id") Long id, Model model) {
+	public String editar(@PathVariable(value="id") Long id, Model model, RedirectAttributes flash) {
 		if (id <= 0) {
+			flash.addFlashAttribute("error", "El ID del cliente no puede ser cero!");
 			return "redirect:/listar";
 		}
 		
 		Cliente cliente = clienteService.findOne(id);
+		if (cliente == null) {
+			flash.addFlashAttribute("error", "El ID del cliente no existe en la BBDD!");
+			return "redirect:/listar";
+		}
 		model.addAttribute("titulo", "Formulario de Cliente");
 		model.addAttribute("cliente", cliente);
 		return "/form";
 	}
 	
 	@PostMapping("/form")
-	public String guardar(@Valid Cliente cliente, BindingResult result, Model model, SessionStatus status) {
+	public String guardar(@Valid Cliente cliente, BindingResult result, Model model, RedirectAttributes flash, SessionStatus status) {
 		if (result.hasErrors()) {
 			model.addAttribute("titulo", "Formulario de Cliente");
 			return "form";
 		}
-		
+		String mensajeFlash = (cliente.getId() != null) ? "Cliente editado con éxito!" : "Cliente creado con éxito!";
+
 		clienteService.save(cliente);
 		status.setComplete();
+		flash.addFlashAttribute("success", mensajeFlash);
 		return "redirect:/listar";
 	}
 	
 	@GetMapping("/eliminar/{id}")
-	public String eliminar(@PathVariable(value="id") Long id) {
-		clienteService.delete(id);
+	public String eliminar(@PathVariable(value="id") Long id, RedirectAttributes flash) {
+		if (id > 0) {
+			clienteService.delete(id);
+			flash.addFlashAttribute("success", "Cliente eliminado con éxito!");
+		}
 		return "redirect:/listar";
 	}
